refactor(db): use bigint from now as we're going to use the smallest possible subunit of currencies (default to USC for now)

+++ db/migrations/000018_create_invoices_table.up.sql
@@ -5,9 +5,9 @@ create table if not exists sales_invoices (
     currency_code text not null references currencies(code),
     inventory_id bigint not null references inventories(id),
     client_id bigint not null references clients(id),
-    amount numeric(12,4) not null,
+    amount bigint not null,
     discount SMALLINT not null CHECK (discount >= 0 AND discount <= 100),
-    net_amount numeric(12,4) not null,
+    net_amount bigint not null,
     created_at timestamp(0) WITH time zone NOT NULL DEFAULT NOW()
 );
